add harness.rs
